Allow dynamic galleries that aren't in the DOM
I ran into an error when using `dynamicEl`:
`TypeError: _this.core.$items.eq is not a function`

In my case `_this.core.$items` is not a jQuery object, but rather an object with `src` and `thumb` values. This broke things in sharer, because it expects $items to be a jQuery object. 

This fixes that error and also separates the data gathering from adding it back to DOM a little bit.